requirements.txt
# Core
fastapi==0.115.12
uvicorn[standard]==0.34.2
python-multipart==0.0.20

# AI / RAG
openai==1.68.2
sentence-transformers==3.4.1
faiss-cpu==1.10.0
numpy==2.2.4

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Database
sqlalchemy==2.0.39
psycopg2-binary==2.9.10

# Config / Utils
python-dotenv==1.0.1
requests==2.32.3
pydantic[email]==2.10.6
